Renamed knowrob_moveit into knowrob_moveit_core to clarify what we're doing.
