+++ knowrob_moveit_core/include/knowrob_moveit_core/knowrob_moveit_core.hpp
@@ -28,9 +28,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef KNOWROB_MOVEIT_KNOWROB_MOVEIT_HPP 
-#define KNOWROB_MOVEIT_KNOWROB_MOVEIT_HPP 
+#ifndef KNOWROB_MOVEIT_CORE_KNOWROB_MOVEIT_CORE_HPP 
+#define KNOWROB_MOVEIT_CORE_KNOWROB_MOVEIT_CORE_HPP 
 
-#include <knowrob_moveit/collision_checking.hpp>
+#include <knowrob_moveit_core/planning_scene.hpp>
 
-#endif // KNOWROB_MOVEIT_KNOWROB_MOVEIT_HPP
+#endif // KNOWROB_MOVEIT_CORE_KNOWROB_MOVEIT_CORE_HPP
