+++ knowrob_moveit_core/launch/planning_scene.launch
@@ -1,0 +1,5 @@
+<launch>
+  <node pkg="knowrob_moveit_core" type="planning_scene" name="planning_scene">
+    <param name="robot_description" textfile="$(find knowrob_moveit_core)/test_data/pr2.urdf" />
+  </node>
+</launch>
