+++ knowrob_moveit_core/test_data/pr2.urdf
